+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @apps/demo
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - react-router-xray@2.0.1
+  - react-router-xray-react@2.0.3
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - react-router-xray-react@2.0.2
+
 ## 0.1.2
 
 ### Patch Changes